+++ packages/docs/docs/main/change-log.md
@@ -26,6 +26,13 @@ This page lists the changes to the IsaacScript framework.
   - `StageAPILayoutRockSubtype`
   - `StageAPILayoutSpikesOnOffVariant`
 - CI in new mods now runs build and lint in a separate jobs so that they run faster.
+- Added the following helper functions:
+  - `shouldWhoreOfBabylonBeActive`
+- Changed the following helper functions:
+  - `spawnCollectible` (is now pure)
+  - `preventCollectibleRotation` (is now pure)
+- Removed the following helper functions:
+  - `spawnCollectibleUnsafe` (use the `spawnCollectible` function instead)
 
 ## November 2nd, 2023